requirements.txt
apache-airflow[crypto,s3,password,postgres]==1.9.0
boto3==1.7.4
psycopg2==2.7.5
requests==2.19.0
flake8==3.5.0
pytest==3.7.0
pytest-cov==2.5.1
wheel==0.31.1